Intial Commit
Implments basic outline of the workflow of the UI. Start Page, User
Login Window, Account Registration Window, Home Page, Opponent Search
Window, and Pictogram Creation Page are all implemented, and running in a
somewhat consistent flow.

The big feature of this commit is the functionality of the Pictogram
Creation (creation page). Currently, we have implemented on the creation
page the ability to upload photos from an outside internet url, display of
that image under "current image", and the ability to "assign" the current
images to one of the thumbnails simply by either clicking on its
respective "assign" button, or by clicking on the thumbnail itself.

The current phrase display on the creation page is simply only text.
Phrase generation as well as user generated image upload are features stil
under development.

NOTE: In its current state the UI is being optimized for communication
with the server, but that communication is not yet functioning. therefore,
no actual user interaction or server interaction is occuring at this point
in time.

tl;dr The app is becoming a reality! Go play with it!